requirements.txt
numpy
opencv-python
pillow
python-multipart
fastapi
uvicorn